reinitialized sentry and moved data to landing
