+++ src/app/(landing)/_data/faqs-data.ts
