+++ src/app/layout.tsx
@@ -5,7 +5,7 @@ import icons from '@/metadata/icons'
 import _isEmpty from 'lodash/isEmpty'
 import robots from '@/metadata/robot'
 import twitter from '@/metadata/twitter'
-import openGraph from '@/metadata/openGraph'
+import openGraph from '@/metadata/open-graph'
 import type { Metadata, Viewport } from 'next'
 import { GoogleAnalytics } from '@next/third-parties/google'
 import { Toaster } from '@/components/ui/toaster'
