--- src/database/index.ts
@@ -1,3 +1,0 @@
-// TODO: Add logger utility function to get rid of direct console.* calls.
-// TODO: Add a utility function to log stacktraces in if NODE_ENV is set to
-//       'development'.
