--- src/database/models/header.ts
@@ -1,9 +1,0 @@
-export interface HeaderProps {
-  missionRef: React.MutableRefObject<null>
-  featuresRef: React.MutableRefObject<null>
-  projectRef: React.MutableRefObject<null>
-  teamRef: React.MutableRefObject<null>
-  signupRef: React.MutableRefObject<null>
-  activePage: string
-  setActivePage: (activePage: string) => void
-}
